Rename: MemberRoleEnum 패키지 이동
