--- src/main/java/com/capstone/rentit/common/MemberRoleEnum.java
@@ -1,5 +1,0 @@
-package com.capstone.rentit.common;
-
-public enum MemberRoleEnum {
-    STUDENT, COUNCIL, COMPANY, ADMIN
-}
