+++ src/main/java/com/capstone/rentit/member/domain/Member.java
@@ -1,6 +1,6 @@
 package com.capstone.rentit.member.domain;
 
-import com.capstone.rentit.common.MemberRoleEnum;
+import com.capstone.rentit.member.status.MemberRoleEnum;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
