+++ src/main/java/com/capstone/rentit/member/service/MemberService.java
@@ -1,6 +1,6 @@
 package com.capstone.rentit.member.service;
 
-import com.capstone.rentit.common.MemberRoleEnum;
+import com.capstone.rentit.member.status.MemberRoleEnum;
 import com.capstone.rentit.member.domain.Company;
 import com.capstone.rentit.member.domain.Member;
 import com.capstone.rentit.member.domain.Student;
