+++ src/test/java/com/capstone/rentit/otp/controller/OtpControllerTest.java
@@ -1,11 +1,8 @@
 package com.capstone.rentit.otp.controller;
 
-import com.capstone.rentit.common.MemberRoleEnum;
+import com.capstone.rentit.member.status.MemberRoleEnum;
 import com.capstone.rentit.config.WebConfig;
-import com.capstone.rentit.exception.GlobalExceptionHandler;
-import com.capstone.rentit.exception.OtpExceptionHandler;
 import com.capstone.rentit.login.dto.MemberDetails;
-import com.capstone.rentit.login.filter.JwtAuthenticationFilter;
 import com.capstone.rentit.login.provider.JwtTokenProvider;
 import com.capstone.rentit.login.service.MemberDetailsService;
 import com.capstone.rentit.member.domain.Student;
